ref(pr-metrics): Extract Seer contract models into contracts.py
The Seer-contract pydantic mirrors lived split across two modules: the
outbound PrCloseJudgeRequest/PrActivityEvent in judge.py and the inbound
PrConversationAnalysis in emit.py. That split made the contract surface hard
to find and kept emit.py importing the conversation model purely so judge.py
could re-import it.

Move all three mirrors — plus the CloseAction vocabulary they share — into a
dedicated contracts.py so the whole Sentry-side mirror of
getsentry/seer:src/seer/pr_metrics/models.py sits in one place. emit.py and
judge.py now both depend on contracts.py, which depends on neither, removing
the would-be import cycle. No behavior change.
